# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Laboria API
  description: API RESTful para plataforma de empleo y formación profesional
  version: 1.0.0
  contact:
    name: Laboria Team
    email: support@laboria.com
    url: https://laboria.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Servidor de desarrollo
  - url: https://api.laboria.com/api
    description: Servidor de producción
paths:
...
  /users/profile:
    get:
      tags:
        - Usuarios
      summary: Obtener perfil de usuario
      description: Obtiene el perfil completo del usuario autenticado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Usuarios
      summary: Actualizar perfil de usuario
      description: Actualiza los datos del perfil del usuario autenticado
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Perfil actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error en la solicitud"
        details:
          type: object
          additionalProperties: true
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operación exitosa"
        data:
          type: object
          nullable: true
...
    UserProfileResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Perfil obtenido exitosamente"
        data:
          type: object
          properties:
            id:
              type: integer
              example: 1
            username:
              type: string
              example: "usuario123"
            email:
              type: string
              format: email
              example: "usuario@ejemplo.com"
            role:
              type: string
              example: "user"
            profile:
              $ref: '#/components/schemas/Profile'
    Profile:
      type: object
      properties:
        nombre:
          type: string
          example: "Juan"
        apellido:
          type: string
          example: "Pérez"
        telefono:
          type: string
          example: "+1234567890"
        bio:
          type: string
          example: "Desarrollador Full Stack con 5 años de experiencia"
        habilidades:
          type: string
          example: "JavaScript, React, Node.js, Python"
        experiencia:
          type: string
          example: "5 años en desarrollo web"
        foto_perfil:
          type: string
          example: "/uploads/avatar.jpg"
        disponibilidad:
          type: string
          example: "inmediata"
        salario_minimo:
          type: number
          example: 50000
        salario_maximo:
          type: number
          example: 80000
    UpdateProfileRequest:
      type: object
      properties:
        nombre:
          type: string
          example: "Juan"
        apellido:
          type: string
          example: "Pérez"
        telefono:
          type: string
          example: "+1234567890"
        bio:
          type: string
          example: "Desarrollador Full Stack con 5 años de experiencia"
        habilidades:
          type: string
          example: "JavaScript, React, Node.js, Python"
        experiencia:
          type: string
          example: "5 años en desarrollo web"
        linkedin:
          type: string
          format: uri
          example: "https://linkedin.com/in/juanperez"
        github:
          type: string
          format: uri
          example: "https://github.com/juanperez"
        website:
          type: string
          format: uri
          example: "https://juanperez.dev"
...